refactor: split Outcome Flow and relocate
